# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bot-Mox API v1
  version: 1.0.0
  description: Canonical API surface for phased backend/frontend alignment.
servers:
  - url: http://localhost:3002
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VmRegisterRequest:
      type: object
      properties:
        vm_uuid:
          type: string
        user_id:
          type: string
        vm_name:
          type: string
        project_id:
          type: string
        status:
          type: string
          enum: [active, paused, revoked]
        metadata:
          type: object
          additionalProperties: true
      required: [vm_uuid]
...
paths:
...
  /api/v1/vm/register:
    post:
      summary: Register VM UUID in tenant-scoped VM registry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VmRegisterRequest'
      responses:
        '201':
          description: VM registered
        '400':
          description: Invalid request body
  /api/v1/vm/{uuid}/resolve:
    get:
      summary: Resolve VM UUID in current tenant
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
      responses:
        '200':
          description: VM registry entry
        '404':
          description: VM UUID not found
...
  /api/v1/vm-ops/proxmox/{action}:
    post:
      summary: Dispatch proxmox command through agent
      parameters:
        - in: path
          name: action
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                agent_id:
                  type: string
                  format: uuid
                action:
                  type: string
                params:
                  type: object
                  additionalProperties: true
              required: [agent_id, action]
      responses:
        '202':
          description: Command queued
        '409':
          description: Agent offline
  /api/v1/vm-ops/syncthing/{action}:
    post:
      summary: Dispatch syncthing command through agent
      parameters:
        - in: path
          name: action
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                agent_id:
                  type: string
                  format: uuid
                action:
                  type: string
                params:
                  type: object
                  additionalProperties: true
              required: [agent_id, action]
      responses:
        '202':
          description: Command queued
        '409':
          description: Agent offline
  /api/v1/vm-ops/commands:
    post:
      summary: Dispatch arbitrary agent command (admin/infra)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                agent_id:
                  type: string
                  format: uuid
                command_type:
                  type: string
                payload:
                  type: object
                  additionalProperties: true
                expires_in_seconds:
                  type: integer
                  default: 300
              required: [agent_id, command_type]
      responses:
        '202':
          description: Command queued
        '409':
          description: Agent offline
    get:
      summary: List agent commands
      parameters:
        - in: query
          name: agent_id
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: Command list
  /api/v1/vm-ops/commands/{id}:
    get:
      summary: Get command status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Command status
        '404':
          description: Not found
    patch:
      summary: Update command status (from agent)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [running, succeeded, failed]
                result:
                  type: object
                  additionalProperties: true
                error_message:
                  type: string
              required: [status]
      responses:
        '200':
          description: Command updated